+++ app/Http/Controllers/DAV/Backend/CalDAV/CalDAVBackend.php
@@ -126,7 +126,7 @@ class CalDAVBackend extends AbstractBackend implements SyncSupport
     {
         $backend = $this->getBackend($calendarId);
         if ($backend) {
-            return $backend->getChanges($syncToken);
+            return $backend->getChanges($calendarId, $syncToken);
         }
     }
 
@@ -165,7 +165,7 @@ class CalDAVBackend extends AbstractBackend implements SyncSupport
     {
         $backend = $this->getBackend($calendarId);
         if ($backend) {
-            $objs = $backend->getObjects();
+            $objs = $backend->getObjects($calendarId);
 
             return $objs
                 ->map(function ($date) use ($backend) {