+++ app/Http/Controllers/DAV/Backend/CalDAV/CalDAVBirthdays.php
@@ -94,10 +94,11 @@ class CalDAVBirthdays extends AbstractCalDAVBackend
     /**
      * Returns the date for the specific uuid.
      *
+     * @param mixed|null $addressBookId
      * @param string  $uuid
      * @return mixed
      */
-    public function getObjectUuid($uuid)
+    public function getObjectUuid($addressBookId, $uuid)
     {
         return SpecialDate::where([
             'account_id' => Auth::user()->account_id,
@@ -110,11 +111,12 @@ class CalDAVBirthdays extends AbstractCalDAVBackend
      *
      * @return \Illuminate\Support\Collection
      */
-    public function getObjects()
+    public function getObjects($addressBookId)
     {
         $contacts = Auth::user()->account
                     ->contacts()
                     ->real()
+                    ->addressBook()
                     ->active()
                     ->get();