+++ app/Http/Controllers/DAV/Backend/CardDAV/CardDAVBackend.php
@@ -4,9 +4,8 @@ namespace App\Http\Controllers\DAV\Backend\CardDAV;
 
 use Sabre\DAV;
 use Illuminate\Support\Arr;
-use App\Models\User\SyncToken;
 use App\Models\Contact\Contact;
-use Sabre\VObject\Component\VCard;
+use App\Models\Account\AddressBook;
 use App\Services\VCard\ExportVCard;
 use App\Services\VCard\ImportVCard;
 use Illuminate\Support\Facades\Log;
@@ -55,11 +54,32 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
      */
     public function getAddressBooksForUser($principalUri)
     {
-        $token = $this->getCurrentSyncToken();
+        $result = [];
+        $result[] = $this->getDefaultAddressBook($principalUri);
+
+        $addressbooks = AddressBook::where('account_id', Auth::user()->account_id)
+            ->get();
+
+        foreach ($addressbooks as $addressbook) {
+            $result[] = [
+                'id'                => $addressbook->addressBookId,
+                'uri'               => $addressbook->addressBookId,
+                'principaluri'      => PrincipalBackend::getPrincipalUser(),
+                '{DAV:}displayname' => $addressbook->name,
+            ];
+        }
+
+        return $result;
+    }
+
+    private function getDefaultAddressBook($principalUri)
+    {
+        $id = $this->backendUri();
+        $token = $this->getCurrentSyncToken(null);
 
         $des = [
-            'id'                => $this->backendUri(),
-            'uri'               => $this->backendUri(),
+            'id'                => $id,
+            'uri'               => $id,
             'principaluri'      => PrincipalBackend::getPrincipalUser(),
             '{DAV:}displayname' => trans('app.dav_contacts'),
             '{'.CardDAVPlugin::NS_CARDDAV.'}addressbook-description' => trans('app.dav_contacts_description', ['name' => Auth::user()->name]),
@@ -79,9 +99,7 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
             ];
         }
 
-        return [
-            $des,
-        ];
+        return $des;
     }
 
     /**
@@ -152,7 +170,7 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
      */
     public function getChangesForAddressBook($addressBookId, $syncToken, $syncLevel, $limit = null)
     {
-        return $this->getChanges($syncToken);
+        return $this->getChanges($addressBookId, $syncToken);
     }
 
     /**
@@ -164,13 +182,16 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
     private function prepareCard($contact)
     {
         try {
-            $vcard = app(ExportVCard::class)
-                ->execute([
-                    'account_id' => Auth::user()->account_id,
-                    'contact_id' => $contact->id,
-                ]);
+            $carddata = $contact->vcard;
+            if (empty($carddat)) {
+                $vcard = app(ExportVCard::class)
+                    ->execute([
+                        'account_id' => Auth::user()->account_id,
+                        'contact_id' => $contact->id,
+                    ]);
 
-            $carddata = $vcard->serialize();
+                $carddata = $vcard->serialize();
+            }
 
             return [
                 'id' => $contact->hashID(),
@@ -188,14 +209,24 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
     /**
      * Returns the contact for the specific uuid.
      *
+     * @param mixed|null $addressBookId
      * @param string  $uuid
      * @return Contact
      */
-    public function getObjectUuid($uuid)
+    public function getObjectUuid($addressBookId, $uuid)
     {
+        $addressBook = null;
+        if ($addressBookId) {
+            $addressBook = AddressBook::where([
+                'account_id' => Auth::user()->account_id,
+                'addressBookId' => $addressBookId == $this->backendUri() ? null : $addressBookId,
+            ])->first();
+        }
+
         return Contact::where([
             'account_id' => Auth::user()->account_id,
             'uuid' => $uuid,
+            'addressbook_id' => $addressBook ? $addressBook->id : null,
         ])->first();
     }
 
@@ -204,11 +235,12 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
      *
      * @return \Illuminate\Support\Collection
      */
-    public function getObjects()
+    public function getObjects($addressBookId)
     {
         return Auth::user()->account
                     ->contacts()
                     ->real()
+                    ->addressBook(Auth::user()->account_id, $addressBookId)
                     ->active()
                     ->get();
     }
@@ -229,12 +261,12 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
      * calculating them. If they are specified, you can also ommit carddata.
      * This may speed up certain requests, especially with large cards.
      *
-     * @param mixed $addressbookId
+     * @param mixed $addressBookId
      * @return array
      */
-    public function getCards($addressbookId)
+    public function getCards($addressBookId)
     {
-        $contacts = $this->getObjects();
+        $contacts = $this->getObjects($addressBookId);
 
         return $contacts->map(function ($contact) {
             return $this->prepareCard($contact);
@@ -255,7 +287,7 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
      */
     public function getCard($addressBookId, $cardUri)
     {
-        $contact = $this->getObject($cardUri);
+        $contact = $this->getObject($addressBookId, $cardUri);
 
         if ($contact) {
             return $this->prepareCard($contact);
@@ -323,7 +355,7 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
     {
         $contact_id = null;
         if ($cardUri) {
-            $contact = $this->getObject($cardUri);
+            $contact = $this->getObject($addressBookId, $cardUri);
 
             if ($contact) {
                 $contact_id = $contact->id;
@@ -338,6 +370,7 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
                     'contact_id' => $contact_id,
                     'entry' => $cardData,
                     'behaviour' => ImportVCard::BEHAVIOUR_REPLACE,
+                    'addressBookId' => $addressBookId == $this->backendUri() ? null : $addressBookId,
                 ]);
 
             if (! Arr::has($result, 'error')) {
@@ -383,8 +416,8 @@ class CardDAVBackend extends AbstractBackend implements SyncSupport, IDAVBackend
      */
     public function updateAddressBook($addressBookId, DAV\PropPatch $propPatch)
     {
-        $propPatch->handle('{'.CalDAVPlugin::NS_CALENDARSERVER.'}me-card', function ($props) {
-            $contact = $this->getObject($props->getHref());
+        $propPatch->handle('{'.CalDAVPlugin::NS_CALENDARSERVER.'}me-card', function ($props) use ($addressBookId) {
+            $contact = $this->getObject($addressBookId, $props->getHref());
 
             $data = [
                 'contact_id' => $contact->id,