+++ app/Http/Controllers/DAV/Backend/IDAVBackend.php
@@ -14,17 +14,18 @@ interface IDAVBackend
     /**
      * Returns the object for the specific uuid.
      *
+     * @param mixed|null $collectionId
      * @param string  $uuid
      * @return mixed
      */
-    public function getObjectUuid($uuid);
+    public function getObjectUuid($collectionId, $uuid);
 
     /**
      * Returns the collection of objects.
      *
      * @return \Illuminate\Support\Collection
      */
-    public function getObjects();
+    public function getObjects($collectionId);
 
     /**
      * Returns the extension for this backend.