+++ database/migrations/2020_03_25_082324_add_contact_address_book_id.php
@@ -1,0 +1,33 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddContactAddressBookId extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('contacts', function (Blueprint $table) {
+            $table->unsignedBigInteger('addressbook_id')->after('account_id')->nullable();
+            $table->foreign('addressbook_id')->references('id')->on('addressbooks')->onDelete('cascade');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('contacts', function (Blueprint $table) {
+            $table->dropColumn('addressbook_id');
+        });
+    }
+}